backend/love_api/api/serializers/todo_calendar_serializer.py
from rest_framework import serializers
from api.models import *

class TodoCalendarSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = TodoCalendar
        exclude = ()